Cleanup todo models
