+++ core/src/main/scala/org/sarden/core/todo/internal/dto.scala
@@ -1,0 +1,12 @@
+package org.sarden.core.todo.internal
+
+import doobie.Read
+
+private[internal] case class TodoDTO(
+    id: String,
+    name: String,
+    schedule: String,
+    notifyBefore: String,
+    lastRun: Option[Long],
+    createdAt: Long,
+) derives Read
