+++ core/src/main/scala/org/sarden/core/todo/internal/instances.scala
@@ -1,0 +1,14 @@
+package org.sarden.core.todo.internal
+
+import io.scalaland.chimney.partial.Result
+import io.scalaland.chimney.{PartialTransformer, Transformer}
+import zio.json.*
+
+import org.sarden.core.todo.TodoSchedule
+
+given JsonCodec[TodoSchedule] = JsonCodec.derived
+
+given PartialTransformer[String, TodoSchedule] = PartialTransformer: raw =>
+  Result.fromEitherString(raw.fromJson)
+
+given Transformer[TodoSchedule, String] = _.toJson
